Updated in to 8.3.7 lates version

+++ src/server/scripts/Draenor/BloodmaulSlagMines/instance_bloodmaul.cpp
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2017-2019 AshamaneProject <https://github.com/AshamaneProject>
- * Copyright (C) 2016 Firestorm Servers <https://firestorm-servers.com>
+ * Copyright (C) 2020 LatinCoreTeam
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -458,11 +457,15 @@ namespace Instances
 
                         if (m_CheckZPosTimer <= diff)
                         {
-                            DoOnPlayers([](Player* player)
+                            Map::PlayerList const& playerList = instance->GetPlayers();
+                            for (Map::PlayerList::const_iterator itr = playerList.begin(); itr != playerList.end(); ++itr)
                             {
-                                if (player->GetPositionZ() <= 150.0f)
-                                    player->Kill(player);
-                            });
+                                if (Player* player = itr->GetSource())
+                                {
+                                    if (player->GetPositionZ() <= 150.0f)
+                                        player->Kill(player);
+                                }
+                            }
 
                             m_CheckZPosTimer = 1000;
                         }
